# src/integration-test/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: care-management
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:63933
  description: Generated server url
...
paths:
...
  /{municipalityId}/{namespace}/errands/financial-assistance/{errandId}/monitorings/{monitoringId}:
    get:
      tags:
      - Financial Assistance Monitorings
      summary: Read a monitoring on an errand
      operationId: getMonitoring
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
      - name: namespace
        in: path
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
      - name: errandId
        in: path
        required: true
        schema:
          type: string
      - name: monitoringId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitoring"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    put:
      tags:
      - Financial Assistance Monitorings
      summary: Replace a monitoring on an errand
      operationId: updateMonitoring
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
      - name: namespace
        in: path
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
      - name: errandId
        in: path
        required: true
        schema:
          type: string
      - name: monitoringId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitoringRequest"
        required: true
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitoring"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
      - Financial Assistance Monitorings
      summary: Delete a monitoring on an errand
      operationId: deleteMonitoring
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
      - name: namespace
        in: path
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
      - name: errandId
        in: path
        required: true
        schema:
          type: string
      - name: monitoringId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Successful operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/{namespace}/errands/financial-assistance/{errandId}/monitorings:
    get:
      tags:
      - Financial Assistance Monitorings
      summary: List the monitorings on an errand
      operationId: listMonitorings
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
      - name: namespace
        in: path
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
      - name: errandId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Monitoring"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
      - Financial Assistance Monitorings
      summary: Create a monitoring on an errand
      description: Creates a date-bound monitoring on the errand. Returns the created
        monitoring and its Location.
      operationId: createMonitoring
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
      - name: namespace
        in: path
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
      - name: errandId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitoringRequest"
        required: true
      responses:
        "201":
          description: Successful operation
          headers:
            Location:
              style: simple
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitoring"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
...
    MonitoringRequest:
      type: object
      description: Request to create or replace an EB monitoring on an errand.
      properties:
        title:
          type: string
          description: Short headline for the monitoring
          examples:
          - Follow up income details from CSN
          minLength: 1
        description:
          type: string
          description: Free-text details of what to watch for
          examples:
          - Awaiting supplementary documentation before a decision can be made.
        startDate:
          type: string
          format: date
          description: When the watch becomes relevant (monitoringsdatum)
          examples:
          - 2026-07-01
        endDate:
          type: string
          format: date
          description: When the watch ends — open-ended when omitted. Must not be
            before the start date.
          examples:
          - 2026-07-31
        createdBy:
          type: string
          description: The caseworker who created the monitoring
          examples:
          - joe01doe
      required:
      - startDate
      - title
    Monitoring:
      type: object
      description: An EB monitoring (date-bound watch/reminder) on an errand.
      properties:
        id:
          type: string
          description: The monitoring id
          examples:
          - f47ac10b-58cc-4372-a567-0e02b2c3d479
          readOnly: true
        title:
          type: string
          description: Short headline for the monitoring
          examples:
          - Follow up income details from CSN
        description:
          type: string
          description: Free-text details of what to watch for
          examples:
          - Awaiting supplementary documentation before a decision can be made.
        startDate:
          type: string
          format: date
          description: When the watch becomes relevant (monitoringsdatum)
          examples:
          - 2026-07-01
        endDate:
          type: string
          format: date
          description: When the watch ends — open-ended when omitted
          examples:
          - 2026-07-31
        createdBy:
          type: string
          description: The caseworker who created the monitoring
          examples:
          - joe01doe
        created:
          type: string
          format: date-time
          description: When the monitoring was created
          readOnly: true
        updated:
          type: string
          format: date-time
          description: When the monitoring was last updated
          readOnly: true
...
  securitySchemes: {}